refactor(core-weave): extract current shape assertions
