+++ src/core/weave/weave.ts
@@ -1,5 +1,4 @@
 import type { Quad } from "n3";
-import * as pathPosix from "@std/path/posix";
 import {
   formatDesignatorPathForDisplay,
   toDesignatorResourcePagePath,
@@ -25,7 +24,6 @@ import {
   type WeaveResourcePageGenerationPolicies,
 } from "./resource_page_policy.ts";
 import {
-  SFCFG_NAMESPACE,
   SFLO_NAMESPACE,
   SFLO_TURTLE_PREFIX_DECLARATION,
 } from "../rdf/namespaces.ts";
@@ -46,32 +44,27 @@ import type {
 import type { RepositorySourceFloatingLocator } from "./source_models.ts";
 import type {
   ExtractionSourceEvidenceModel,
-  PayloadWorkingArtifact,
-  ReferenceTargetSourcePayloadArtifact,
   ResourcePageDefinitionWorkingArtifact,
   WeaveableKnopCandidate,
 } from "./candidates.ts";
 import type { PlanWeaveInput, WeavePlan } from "./planning_models.ts";
+import type {
+  MeshInventoryProgression,
+  PageDefinitionWeaveProgression,
+} from "./progression_models.ts";
 import type { WeaveSlice } from "./slices.ts";
 import { toArtifactManifestationPath } from "./artifact_manifestation_paths.ts";
-import { isDeclaredArtifactHistory } from "./artifact_history_queries.ts";
 import {
   assertOverwriteExistingStateTargets,
   planOverwriteExistingPayloadState,
 } from "./payload_overwrite.ts";
 import {
   type PayloadVersionLayout,
-  requirePayloadCurrentStatePath,
-  requirePayloadHistoryPath,
   resolveFirstPayloadVersionLayout,
   resolveSecondPayloadVersionLayout,
 } from "./payload_version_layout.ts";
 import {
-  hasLiteralFact,
   hasNamedNodeFact,
-  hasPredicateFact,
-  hasSubjectPredicateFact,
-  hasTermKeyNamedNodeFact,
   parseWeaveShapeQuads,
   requireLiteralValue,
   requireNamedNodePath,
@@ -82,11 +75,8 @@ import {
   resolveOptionalLiteralObject,
   resolveOptionalNamedNodePath,
   resolveOptionalNonNegativeIntegerLiteral,
-  resolveOptionalSegmentHint,
-  resolveUniqueLiteralValuesForTermKey,
   toAbsoluteIri,
   toMeshRelativePath,
-  toRdfTermKey,
 } from "./rdf_helpers.ts";
 import {
   appendPredicateToSubjectBlock,
@@ -100,6 +90,25 @@ import {
   upsertSubjectBlockAfter,
 } from "./turtle_blocks.ts";
 import { classifyWeaveSlice } from "./slice_classification.ts";
+import { assertHasCurrentWorkingFileLocator } from "./source_locator_assertions.ts";
+import {
+  assertCurrentKnopInventoryBaseShape,
+  assertCurrentKnopInventoryShapeForFirstExtractedKnopWeave,
+  assertCurrentKnopInventoryShapeForFirstPageDefinitionWeave,
+  assertCurrentKnopInventoryShapeForFirstReferenceCatalogWeave,
+  assertCurrentKnopInventoryShapeForSecondPayloadWeave,
+  assertCurrentKnopInventoryShapeForSubsequentPageDefinitionWeave,
+  assertCurrentKnopInventoryWithoutHistory,
+  assertCurrentKnopMetadataShape,
+  assertCurrentMeshInventoryShapeForFirstExtractedKnopWeave,
+  assertCurrentMeshInventoryShapeForFirstReferenceCatalogWeave,
+  assertCurrentPayloadArtifactShape,
+  assertCurrentSourceRegistryShapeForFirstExtractedKnopWeave,
+  assertHasNamedNodeFacts,
+  assertReferenceTargetSourcePayloadShapeForFirstExtractedKnopWeave,
+  resolveMeshInventoryProgressionFromMetadata,
+} from "./shape_assertions.ts";
+import { usesMeshLocalWorkingLocatedFile } from "./working_file_paths.ts";
 
 export { WeaveInputError } from "./errors.ts";
 export { planMeshSupportResourcePages } from "./mesh_support_pages.ts";
@@ -158,19 +167,7 @@ export type { WeaveSlice } from "./slices.ts";
 export type { VersionPlan } from "./version_plan.ts";
 
 const RDF_TYPE_IRI = "http://www.w3.org/1999/02/22-rdf-syntax-ns#type";
-const XSD_NON_NEGATIVE_INTEGER_IRI =
-  "http://www.w3.org/2001/XMLSchema#nonNegativeInteger";
-const SFCFG_HAS_NEXT_STATE_SEGMENT_HINT_IRI =
-  `${SFCFG_NAMESPACE}hasNextStateSegmentHint`;
-const SFLO_ARTIFACT_RESOLUTION_MODE_WORKING_IRI =
-  `${SFLO_NAMESPACE}artifactResolutionMode_working`;
-const SFLO_EXTRACTION_SOURCE_IRI = `${SFLO_NAMESPACE}ExtractionSource`;
-const SFLO_HAS_ARTIFACT_RESOLUTION_MODE_IRI =
-  `${SFLO_NAMESPACE}hasArtifactResolutionMode`;
 const SFLO_HAS_EXTRACTION_SOURCE_IRI = `${SFLO_NAMESPACE}hasExtractionSource`;
-const SFLO_HAS_REQUESTED_TARGET_STATE_IRI =
-  `${SFLO_NAMESPACE}hasRequestedTargetState`;
-const SFLO_HAS_TARGET_ARTIFACT_IRI = `${SFLO_NAMESPACE}hasTargetArtifact`;
 const SFLO_HAS_KNOP_ASSET_BUNDLE_IRI = `${SFLO_NAMESPACE}hasKnopAssetBundle`;
 const SFLO_HAS_RESOURCE_PAGE_DEFINITION_IRI =
   `${SFLO_NAMESPACE}hasResourcePageDefinition`;
@@ -179,7 +176,6 @@ const SFLO_RESOURCE_PAGE_DEFINITION_IRI =
   `${SFLO_NAMESPACE}ResourcePageDefinition`;
 const SFLO_CURRENT_ARTIFACT_HISTORY_IRI =
   `${SFLO_NAMESPACE}currentArtifactHistory`;
-const SFLO_DESIGNATOR_PATH_IRI = `${SFLO_NAMESPACE}designatorPath`;
 const SFLO_DIGITAL_ARTIFACT_IRI = `${SFLO_NAMESPACE}DigitalArtifact`;
 const SFLO_HAS_ARTIFACT_HISTORY_IRI = `${SFLO_NAMESPACE}hasArtifactHistory`;
 const SFLO_HAS_MANIFESTATION_IRI = `${SFLO_NAMESPACE}hasManifestation`;
@@ -189,33 +185,21 @@ const SFLO_HAS_KNOP_INVENTORY_IRI = `${SFLO_NAMESPACE}hasKnopInventory`;
 const SFLO_HAS_KNOP_METADATA_IRI = `${SFLO_NAMESPACE}hasKnopMetadata`;
 const SFLO_HAS_KNOP_SOURCE_REGISTRY_IRI =
   `${SFLO_NAMESPACE}hasKnopSourceRegistry`;
-const SFLO_HAS_PAYLOAD_ARTIFACT_IRI = `${SFLO_NAMESPACE}hasPayloadArtifact`;
 const SFLO_HAS_REFERENCE_CATALOG_IRI = `${SFLO_NAMESPACE}hasReferenceCatalog`;
 const SFLO_HAS_REFERENCE_LINK_IRI = `${SFLO_NAMESPACE}hasReferenceLink`;
 const SFLO_HAS_REFERENCE_ROLE_IRI = `${SFLO_NAMESPACE}hasReferenceRole`;
 const SFLO_HAS_RESOURCE_PAGE_IRI = `${SFLO_NAMESPACE}hasResourcePage`;
-const SFLO_HAS_SOURCE_BINDING_IRI = `${SFLO_NAMESPACE}hasSourceBinding`;
 const SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI =
   `${SFLO_NAMESPACE}hasWorkingKnopInventoryFile`;
 const SFLO_HAS_WORKING_LOCATED_FILE_IRI =
   `${SFLO_NAMESPACE}hasWorkingLocatedFile`;
-const SFLO_HAS_REPOSITORY_SOURCE_FLOATING_LOCATOR_IRI =
-  `${SFLO_NAMESPACE}hasRepositorySourceFloatingLocator`;
-const SFLO_REPOSITORY_SOURCE_FLOATING_LOCATOR_IRI =
-  `${SFLO_NAMESPACE}RepositorySourceFloatingLocator`;
-const SFLO_SOURCE_REPOSITORY_URL_IRI = `${SFLO_NAMESPACE}sourceRepositoryUrl`;
-const SFLO_SOURCE_REPOSITORY_PATH_FROM_ROOT_IRI =
-  `${SFLO_NAMESPACE}sourceRepositoryPathFromRoot`;
 const SFLO_WORKING_FILE_PATH_IRI = `${SFLO_NAMESPACE}workingLocalRelativePath`;
 const SFLO_KNOP_IRI = `${SFLO_NAMESPACE}Knop`;
 const SFLO_KNOP_INVENTORY_IRI = `${SFLO_NAMESPACE}KnopInventory`;
-const SFLO_KNOP_METADATA_IRI = `${SFLO_NAMESPACE}KnopMetadata`;
 const SFLO_KNOP_SOURCE_REGISTRY_IRI = `${SFLO_NAMESPACE}KnopSourceRegistry`;
 const SFLO_LATEST_HISTORICAL_STATE_IRI =
   `${SFLO_NAMESPACE}latestHistoricalState`;
-const SFLO_LOCATED_FILE_IRI = `${SFLO_NAMESPACE}LocatedFile`;
 const SFLO_MESH_INVENTORY_IRI = `${SFLO_NAMESPACE}MeshInventory`;
-const SFLO_NEXT_HISTORY_ORDINAL_IRI = `${SFLO_NAMESPACE}nextHistoryOrdinal`;
 const SFLO_NEXT_STATE_ORDINAL_IRI = `${SFLO_NAMESPACE}nextStateOrdinal`;
 const SFLO_PAYLOAD_ARTIFACT_IRI = `${SFLO_NAMESPACE}PayloadArtifact`;
 const SFLO_RDF_DOCUMENT_IRI = `${SFLO_NAMESPACE}RdfDocument`;
@@ -228,26 +212,6 @@ const SFLO_REFERENCE_TARGET_STATE_IRI = `${SFLO_NAMESPACE}referenceTargetState`;
 interface SelectedWeaveableKnopCandidate {
   candidate: WeaveableKnopCandidate;
   target?: NormalizedVersionTargetSpec;
-}
-
-interface MeshInventoryProgression {
-  historyPath: string;
-  nextHistoryOrdinal?: number;
-  latestStatePath: string;
-  latestStateOrdinal: number;
-  latestManifestationPath: string;
-  nextStatePath: string;
-  nextStateOrdinal: number;
-}
-
-interface PageDefinitionWeaveProgression {
-  historyPath: string;
-  latestStatePath?: string;
-  latestStateOrdinal: number;
-  nextStatePath: string;
-  nextStateOrdinal: number;
-  nextManifestationPath: string;
-  nextSnapshotPath: string;
 }
 
 export function planWeave(input: PlanWeaveInput): WeavePlan {
@@ -1600,80 +1564,6 @@ function resolveCurrentMeshInventoryProgressionForFirstKnopWeave(
   return progression;
 }
 
-function resolveMeshInventoryProgressionFromMetadata(
-  meshBase: string,
-  currentMeshMetadataTurtle: string | undefined,
-  errorMessage: string,
-): MeshInventoryProgression {
-  if (currentMeshMetadataTurtle === undefined) {
-    throw new WeaveInputError(errorMessage);
-  }
-
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentMeshMetadataTurtle,
-    errorMessage,
-  );
-  const meshInventoryIri = toAbsoluteIri(meshBase, "_mesh/_inventory");
-  const historyIri = requireSingleNamedNodeObject(
-    quads,
-    meshInventoryIri,
-    SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-    errorMessage,
-  );
-  const nextHistoryOrdinal = requireSingleNonNegativeIntegerLiteral(
-    quads,
-    meshInventoryIri,
-    SFLO_NEXT_HISTORY_ORDINAL_IRI,
-    errorMessage,
-  );
-  const historyPath = toMeshRelativePath(
-    meshBase,
-    historyIri,
-    "the current MeshInventory history",
-  );
-  const latestStateIri = requireSingleNamedNodeObject(
-    quads,
-    historyIri,
-    SFLO_LATEST_HISTORICAL_STATE_IRI,
-    errorMessage,
-  );
-  const latestStatePath = toMeshRelativePath(
-    meshBase,
-    latestStateIri,
-    "the latest MeshInventory historical state",
-  );
-  const nextStateOrdinal = requireSingleNonNegativeIntegerLiteral(
-    quads,
-    historyIri,
-    SFLO_NEXT_STATE_ORDINAL_IRI,
-    errorMessage,
-  );
-  if (nextStateOrdinal === 0) {
-    throw new WeaveInputError(errorMessage);
-  }
-  const latestStateOrdinal = nextStateOrdinal - 1;
-  const nextStateSegmentHint = resolveOptionalSegmentHint(
-    quads,
-    historyIri,
-    SFCFG_HAS_NEXT_STATE_SEGMENT_HINT_IRI,
-    errorMessage,
-  );
-  const nextStatePath = `${historyPath}/${
-    nextStateSegmentHint ?? toStateSegment(nextStateOrdinal)
-  }`;
-
-  return {
-    historyPath,
-    nextHistoryOrdinal,
-    latestStatePath,
-    latestStateOrdinal,
-    latestManifestationPath: `${latestStatePath}/ttl`,
-    nextStatePath,
-    nextStateOrdinal,
-  };
-}
-
 function resolveCurrentMeshInventoryProgressionForFirstPayloadWeave(
   meshBase: string,
   currentMeshInventoryTurtle: string,
@@ -1746,873 +1636,6 @@ function resolveCurrentMeshInventoryProgressionForFirstPayloadWeave(
     ...progression,
     latestManifestationPath,
   };
-}
-
-function assertCurrentMeshInventoryShapeForFirstReferenceCatalogWeave(
-  meshBase: string,
-  currentMeshInventoryTurtle: string,
-  currentMeshMetadataTurtle: string | undefined,
-  designatorPath: string,
-): void {
-  const knopPath = toKnopPath(designatorPath);
-  const errorMessage =
-    "The current local weave slice only supports the settled 08 pre-weave reference-catalog mesh inventory shape.";
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentMeshInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    ["_mesh/_inventory", RDF_TYPE_IRI, SFLO_MESH_INVENTORY_IRI],
-    ["_mesh/_inventory", RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    ["_mesh/_inventory", RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-  ]);
-
-  const historyIri = requireOptionalNamedNodeObject(
-    quads,
-    toAbsoluteIri(meshBase, "_mesh/_inventory"),
-    SFLO_HAS_ARTIFACT_HISTORY_IRI,
-    errorMessage,
-  );
-  if (historyIri === undefined) {
-    return;
-  }
-
-  const progression = resolveMeshInventoryProgressionFromMetadata(
-    meshBase,
-    currentMeshMetadataTurtle,
-    errorMessage,
-  );
-  const historyPath = toMeshRelativePath(
-    meshBase,
-    historyIri,
-    "the current MeshInventory history",
-  );
-  if (
-    progression.historyPath !== historyPath ||
-    toHistoryPathFromStatePath(progression.latestStatePath) !==
-      progression.historyPath ||
-    progression.nextStateOrdinal !== progression.latestStateOrdinal + 1
-  ) {
-    throw new WeaveInputError(errorMessage);
-  }
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [
-      "_mesh/_inventory",
-      SFLO_HAS_ARTIFACT_HISTORY_IRI,
-      progression.historyPath,
-    ],
-    [
-      progression.historyPath,
-      SFLO_HAS_HISTORICAL_STATE_IRI,
-      progression.latestStatePath,
-    ],
-  ]);
-}
-
-function assertCurrentMeshInventoryShapeForFirstExtractedKnopWeave(
-  meshBase: string,
-  currentMeshInventoryTurtle: string,
-  meshInventoryProgression: MeshInventoryProgression | undefined,
-  designatorPath: string,
-  sourcePayloadDesignatorPath: string,
-  sourcePayloadArtifact: Pick<
-    ReferenceTargetSourcePayloadArtifact,
-    "workingLocalRelativePath" | "repositorySourceFloatingLocator"
-  >,
-): void {
-  const rootDesignatorPath = toRootDesignatorPath(sourcePayloadDesignatorPath);
-  const rootKnopPath = toKnopPath(rootDesignatorPath);
-  const sourceKnopPath = toKnopPath(sourcePayloadDesignatorPath);
-  const knopPath = toKnopPath(designatorPath);
-  const designatorPagePath = toDesignatorResourcePagePath(designatorPath);
-  const sourcePayloadPagePath = toDesignatorResourcePagePath(
-    sourcePayloadDesignatorPath,
-  );
-  const errorMessage =
-    `The current local weave slice only supports the settled extracted-knop pre-weave mesh inventory shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentMeshInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    ["_mesh/_inventory", RDF_TYPE_IRI, SFLO_MESH_INVENTORY_IRI],
-    ["_mesh/_inventory", RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    ["_mesh/_inventory", RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    ...(meshInventoryProgression === undefined ? [] : [
-      [
-        "_mesh/_inventory",
-        SFLO_HAS_ARTIFACT_HISTORY_IRI,
-        meshInventoryProgression.historyPath,
-      ],
-      [
-        meshInventoryProgression.historyPath,
-        SFLO_HAS_HISTORICAL_STATE_IRI,
-        meshInventoryProgression.latestStatePath,
-      ],
-    ] as const),
-    [sourcePayloadDesignatorPath, RDF_TYPE_IRI, SFLO_PAYLOAD_ARTIFACT_IRI],
-    [sourcePayloadDesignatorPath, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [sourcePayloadDesignatorPath, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [
-      sourcePayloadDesignatorPath,
-      SFLO_HAS_RESOURCE_PAGE_IRI,
-      sourcePayloadPagePath,
-    ],
-    [rootKnopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [
-      rootKnopPath,
-      SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI,
-      `${rootKnopPath}/_inventory/inventory.ttl`,
-    ],
-    [
-      rootKnopPath,
-      SFLO_HAS_RESOURCE_PAGE_IRI,
-      `${rootKnopPath}/index.html`,
-    ],
-    [sourceKnopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [
-      sourceKnopPath,
-      SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI,
-      `${sourceKnopPath}/_inventory/inventory.ttl`,
-    ],
-    [
-      sourceKnopPath,
-      SFLO_HAS_RESOURCE_PAGE_IRI,
-      `${sourceKnopPath}/index.html`,
-    ],
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-  ]);
-  if (meshInventoryProgression !== undefined) {
-    if (
-      meshInventoryProgression.historyPath !== "_mesh/_inventory/_history001" ||
-      toHistoryPathFromStatePath(meshInventoryProgression.latestStatePath) !==
-        meshInventoryProgression.historyPath ||
-      meshInventoryProgression.nextStateOrdinal !==
-        meshInventoryProgression.latestStateOrdinal + 1
-    ) {
-      throw new WeaveInputError(errorMessage);
-    }
-  }
-  assertHasCurrentSourceLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    sourcePayloadDesignatorPath,
-    sourcePayloadArtifact,
-  );
-
-  if (
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      designatorPath,
-      SFLO_HAS_RESOURCE_PAGE_IRI,
-      designatorPagePath,
-    ) ||
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      knopPath,
-      SFLO_HAS_RESOURCE_PAGE_IRI,
-      `${knopPath}/index.html`,
-    )
-  ) {
-    throw new WeaveInputError(
-      `Mesh inventory already exposes current woven pages for ${designatorPath}.`,
-    );
-  }
-}
-
-function assertCurrentKnopInventoryShapeForFirstExtractedKnopWeave(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  designatorPath: string,
-): void {
-  const knopPath = toKnopPath(designatorPath);
-  const sourceRegistryPath = `${knopPath}/_sources`;
-  const sourcesFilePath = `${sourceRegistryPath}/sources.ttl`;
-  const extractionSourcePath = `${sourceRegistryPath}#extraction-source`;
-  const errorMessage =
-    `The current local weave slice only supports the settled extracted-knop inventory shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [knopPath, SFLO_HAS_KNOP_METADATA_IRI, `${knopPath}/_meta`],
-    [knopPath, SFLO_HAS_KNOP_INVENTORY_IRI, `${knopPath}/_inventory`],
-    [
-      knopPath,
-      SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI,
-      `${knopPath}/_inventory/inventory.ttl`,
-    ],
-    [`${knopPath}/_meta`, RDF_TYPE_IRI, SFLO_KNOP_METADATA_IRI],
-    [`${knopPath}/_meta`, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [`${knopPath}/_meta`, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_KNOP_INVENTORY_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [knopPath, SFLO_HAS_KNOP_SOURCE_REGISTRY_IRI, sourceRegistryPath],
-    [knopPath, SFLO_HAS_EXTRACTION_SOURCE_IRI, extractionSourcePath],
-    [sourceRegistryPath, RDF_TYPE_IRI, SFLO_KNOP_SOURCE_REGISTRY_IRI],
-    [
-      sourceRegistryPath,
-      SFLO_HAS_WORKING_LOCATED_FILE_IRI,
-      sourcesFilePath,
-    ],
-    [sourcesFilePath, RDF_TYPE_IRI, SFLO_LOCATED_FILE_IRI],
-    [sourcesFilePath, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-  ]);
-  assertHasCurrentWorkingFileLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    `${knopPath}/_meta`,
-    `${knopPath}/_meta/meta.ttl`,
-  );
-  assertHasCurrentWorkingFileLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    `${knopPath}/_inventory`,
-    `${knopPath}/_inventory/inventory.ttl`,
-  );
-
-  if (hasPredicateFact(quads, SFLO_HAS_ARTIFACT_HISTORY_IRI)) {
-    throw new WeaveInputError(
-      `Extracted KnopInventory already has explicit woven history for ${designatorPath}.`,
-    );
-  }
-}
-
-function assertCurrentSourceRegistryShapeForFirstExtractedKnopWeave(
-  meshBase: string,
-  currentSourceRegistryTurtle: string,
-  designatorPath: string,
-  sourceDesignatorPath: string,
-  sourceStatePath: string,
-): void {
-  const sourceRegistryPath = `${toKnopPath(designatorPath)}/_sources`;
-  const extractionSourcePath = `${sourceRegistryPath}#extraction-source`;
-  const errorMessage =
-    `The current local weave slice only supports the settled extracted-knop source registry shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentSourceRegistryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [sourceRegistryPath, RDF_TYPE_IRI, SFLO_KNOP_SOURCE_REGISTRY_IRI],
-    [sourceRegistryPath, SFLO_HAS_SOURCE_BINDING_IRI, extractionSourcePath],
-    [extractionSourcePath, RDF_TYPE_IRI, SFLO_EXTRACTION_SOURCE_IRI],
-    [extractionSourcePath, SFLO_HAS_TARGET_ARTIFACT_IRI, sourceDesignatorPath],
-  ]);
-  const hasWorkingResolutionMode = hasNamedNodeFact(
-    quads,
-    meshBase,
-    extractionSourcePath,
-    SFLO_HAS_ARTIFACT_RESOLUTION_MODE_IRI,
-    SFLO_ARTIFACT_RESOLUTION_MODE_WORKING_IRI,
-  );
-  const hasExactTargetState = hasNamedNodeFact(
-    quads,
-    meshBase,
-    extractionSourcePath,
-    SFLO_HAS_REQUESTED_TARGET_STATE_IRI,
-    sourceStatePath,
-  );
-  if (!hasWorkingResolutionMode && !hasExactTargetState) {
-    throw new WeaveInputError(errorMessage);
-  }
-}
-
-function assertReferenceTargetSourcePayloadShapeForFirstExtractedKnopWeave(
-  payloadArtifact: ReferenceTargetSourcePayloadArtifact,
-): void {
-  if (!payloadArtifact.workingLocalRelativePath.endsWith(".ttl")) {
-    throw new WeaveInputError(
-      `The current local extracted weave slice only supports Turtle source payload files; found ${payloadArtifact.workingLocalRelativePath}.`,
-    );
-  }
-
-  const expectedPrefix = payloadArtifact.designatorPath.length === 0
-    ? undefined
-    : `${payloadArtifact.designatorPath}/`;
-  if (
-    payloadArtifact.latestHistoricalStatePath.length === 0 ||
-    (expectedPrefix !== undefined &&
-      !payloadArtifact.latestHistoricalStatePath.startsWith(expectedPrefix))
-  ) {
-    throw new WeaveInputError(
-      `The current local extracted weave slice only supports source payloads with an explicit current historical state.`,
-    );
-  }
-}
-
-function assertCurrentKnopMetadataShape(
-  meshBase: string,
-  currentKnopMetadataTurtle: string,
-  designatorPath: string,
-  knopPath: string,
-): void {
-  const errorMessage =
-    `The current local weave slice only supports the settled first-history KnopMetadata shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopMetadataTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [
-      knopPath,
-      SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI,
-      `${knopPath}/_inventory/inventory.ttl`,
-    ],
-  ]);
-  assertHasLiteralFacts(quads, meshBase, errorMessage, [
-    [knopPath, SFLO_DESIGNATOR_PATH_IRI, designatorPath],
-  ]);
-
-  if (hasPredicateFact(quads, SFLO_HAS_ARTIFACT_HISTORY_IRI)) {
-    throw new WeaveInputError(
-      `KnopMetadata already has explicit history for ${designatorPath}.`,
-    );
-  }
-}
-
-function assertCurrentKnopInventoryBaseShape(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  knopPath: string,
-): void {
-  const errorMessage =
-    `The current local weave slice only supports the settled first-history KnopInventory shape for ${knopPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [knopPath, SFLO_HAS_KNOP_METADATA_IRI, `${knopPath}/_meta`],
-    [knopPath, SFLO_HAS_KNOP_INVENTORY_IRI, `${knopPath}/_inventory`],
-    [
-      knopPath,
-      SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI,
-      `${knopPath}/_inventory/inventory.ttl`,
-    ],
-  ]);
-}
-
-function assertCurrentKnopInventoryWithoutHistory(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  knopPath: string,
-): void {
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    `Could not parse the current KnopInventory while checking current history state for ${knopPath}.`,
-  );
-
-  if (
-    hasSubjectPredicateFact(
-      quads,
-      meshBase,
-      `${knopPath}/_inventory`,
-      SFLO_HAS_ARTIFACT_HISTORY_IRI,
-    )
-  ) {
-    throw new WeaveInputError(
-      `KnopInventory already has explicit history for ${knopPath}.`,
-    );
-  }
-}
-
-function assertCurrentPayloadArtifactShape(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  designatorPath: string,
-  payloadArtifact: PayloadWorkingArtifact,
-): void {
-  const errorMessage =
-    `The current local weave slice only supports the settled integrated payload shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [designatorPath, RDF_TYPE_IRI, SFLO_PAYLOAD_ARTIFACT_IRI],
-    [designatorPath, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [designatorPath, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-  ]);
-  assertHasCurrentPayloadSourceLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    designatorPath,
-    payloadArtifact,
-  );
-
-  const existingHistoryPaths = resolveNamedNodeObjectPaths(
-    quads,
-    meshBase,
-    designatorPath,
-    SFLO_HAS_ARTIFACT_HISTORY_IRI,
-    errorMessage,
-  );
-  const currentHistoryPath = resolveOptionalNamedNodePath(
-    quads,
-    meshBase,
-    designatorPath,
-    SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-    errorMessage,
-  );
-  const hasExistingDeclaredHistory = existingHistoryPaths.some((path) =>
-    isDeclaredArtifactHistory(quads, meshBase, path)
-  );
-  if (
-    hasExistingDeclaredHistory ||
-    (currentHistoryPath !== undefined &&
-      isDeclaredArtifactHistory(quads, meshBase, currentHistoryPath))
-  ) {
-    throw new WeaveInputError(
-      `Payload artifact already has explicit history for ${designatorPath}.`,
-    );
-  }
-}
-
-function assertCurrentKnopInventoryShapeForFirstPageDefinitionWeave(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  designatorPath: string,
-  pageDefinitionArtifact: ResourcePageDefinitionWorkingArtifact,
-  knopInventoryProgression: MeshInventoryProgression,
-): void {
-  const knopPath = toKnopPath(designatorPath);
-  const errorMessage =
-    `The current local weave slice only supports the settled page-definition inventory shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [knopPath, SFLO_HAS_KNOP_METADATA_IRI, `${knopPath}/_meta`],
-    [knopPath, SFLO_HAS_KNOP_INVENTORY_IRI, `${knopPath}/_inventory`],
-    [knopPath, SFLO_HAS_RESOURCE_PAGE_DEFINITION_IRI, `${knopPath}/_page`],
-    [
-      knopPath,
-      SFLO_HAS_WORKING_KNOP_INVENTORY_FILE_IRI,
-      `${knopPath}/_inventory/inventory.ttl`,
-    ],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_KNOP_INVENTORY_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [
-      `${knopPath}/_inventory`,
-      SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-      `${knopPath}/_inventory/_history001`,
-    ],
-    [`${knopPath}/_page`, RDF_TYPE_IRI, SFLO_RESOURCE_PAGE_DEFINITION_IRI],
-    [`${knopPath}/_page`, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [`${knopPath}/_page`, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-  ]);
-  assertHasCurrentWorkingFileLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    `${knopPath}/_page`,
-    pageDefinitionArtifact.workingLocalRelativePath,
-  );
-  assertHasLiteralFacts(quads, meshBase, errorMessage, [[
-    `${knopPath}/_inventory/_history001`,
-    SFLO_NEXT_STATE_ORDINAL_IRI,
-    String(knopInventoryProgression.nextStateOrdinal),
-    XSD_NON_NEGATIVE_INTEGER_IRI,
-  ]]);
-
-  if (
-    !hasNamedNodeFact(
-      quads,
-      meshBase,
-      `${knopPath}/_inventory/_history001`,
-      SFLO_LATEST_HISTORICAL_STATE_IRI,
-      knopInventoryProgression.latestStatePath,
-    )
-  ) {
-    throw new WeaveInputError(
-      `KnopInventory is not at the expected pre-page-definition history state for ${designatorPath}.`,
-    );
-  }
-
-  if (
-    resolveOptionalNamedNodePath(
-      quads,
-      meshBase,
-      `${knopPath}/_page`,
-      SFLO_HAS_ARTIFACT_HISTORY_IRI,
-      errorMessage,
-    ) ||
-    resolveOptionalNamedNodePath(
-      quads,
-      meshBase,
-      `${knopPath}/_page`,
-      SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-      errorMessage,
-    )
-  ) {
-    throw new WeaveInputError(
-      `ResourcePageDefinition already has explicit history for ${designatorPath}.`,
-    );
-  }
-
-  if (pageDefinitionArtifact.assetBundlePath) {
-    assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-      [
-        knopPath,
-        SFLO_HAS_KNOP_ASSET_BUNDLE_IRI,
-        pageDefinitionArtifact.assetBundlePath,
-      ],
-      [
-        pageDefinitionArtifact.assetBundlePath,
-        RDF_TYPE_IRI,
-        SFLO_KNOP_ASSET_BUNDLE_IRI,
-      ],
-    ]);
-  }
-}
-
-function assertCurrentKnopInventoryShapeForSubsequentPageDefinitionWeave(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  designatorPath: string,
-  pageDefinitionArtifact: ResourcePageDefinitionWorkingArtifact,
-  progression: PageDefinitionWeaveProgression,
-  knopInventoryProgression: MeshInventoryProgression,
-): void {
-  const knopPath = toKnopPath(designatorPath);
-  const errorMessage =
-    `The current local weave slice only supports the settled later page-definition inventory shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [knopPath, SFLO_HAS_KNOP_METADATA_IRI, `${knopPath}/_meta`],
-    [knopPath, SFLO_HAS_KNOP_INVENTORY_IRI, `${knopPath}/_inventory`],
-    [knopPath, SFLO_HAS_REFERENCE_CATALOG_IRI, `${knopPath}/_references`],
-    [knopPath, SFLO_HAS_RESOURCE_PAGE_DEFINITION_IRI, `${knopPath}/_page`],
-    [
-      `${knopPath}/_inventory`,
-      SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-      knopInventoryProgression.historyPath,
-    ],
-    [
-      `${knopPath}/_inventory`,
-      SFLO_HAS_WORKING_LOCATED_FILE_IRI,
-      `${knopPath}/_inventory/inventory.ttl`,
-    ],
-    [
-      `${knopPath}/_page`,
-      SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-      progression.historyPath,
-    ],
-    [
-      progression.historyPath,
-      SFLO_LATEST_HISTORICAL_STATE_IRI,
-      progression.latestStatePath!,
-    ],
-  ]);
-  assertHasCurrentWorkingFileLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    `${knopPath}/_page`,
-    pageDefinitionArtifact.workingLocalRelativePath,
-  );
-  assertHasLiteralFacts(quads, meshBase, errorMessage, [
-    [
-      progression.historyPath,
-      SFLO_NEXT_STATE_ORDINAL_IRI,
-      String(progression.nextStateOrdinal),
-      XSD_NON_NEGATIVE_INTEGER_IRI,
-    ],
-    [
-      knopInventoryProgression.historyPath,
-      SFLO_NEXT_STATE_ORDINAL_IRI,
-      String(knopInventoryProgression.nextStateOrdinal),
-      XSD_NON_NEGATIVE_INTEGER_IRI,
-    ],
-  ]);
-
-  if (
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      progression.historyPath,
-      SFLO_HAS_HISTORICAL_STATE_IRI,
-      progression.nextStatePath,
-    ) ||
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      knopInventoryProgression.historyPath,
-      SFLO_HAS_HISTORICAL_STATE_IRI,
-      knopInventoryProgression.nextStatePath,
-    )
-  ) {
-    throw new WeaveInputError(
-      `ResourcePageDefinition already has a later explicit historical state for ${designatorPath}.`,
-    );
-  }
-
-  if (pageDefinitionArtifact.assetBundlePath) {
-    assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-      [
-        knopPath,
-        SFLO_HAS_KNOP_ASSET_BUNDLE_IRI,
-        pageDefinitionArtifact.assetBundlePath,
-      ],
-      [
-        pageDefinitionArtifact.assetBundlePath,
-        RDF_TYPE_IRI,
-        SFLO_KNOP_ASSET_BUNDLE_IRI,
-      ],
-    ]);
-  }
-}
-
-function assertCurrentKnopInventoryShapeForFirstReferenceCatalogWeave(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  designatorPath: string,
-  workingLocalRelativePath: string,
-  options?: { versionKnopInventory?: boolean },
-): void {
-  const knopPath = toKnopPath(designatorPath);
-  const referenceCatalogPath = `${knopPath}/_references`;
-  const versionKnopInventory = options?.versionKnopInventory ?? true;
-  const errorMessage =
-    `The current local weave slice only supports the settled first ReferenceCatalog weave shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [knopPath, SFLO_HAS_REFERENCE_CATALOG_IRI, referenceCatalogPath],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_KNOP_INVENTORY_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [referenceCatalogPath, RDF_TYPE_IRI, SFLO_REFERENCE_CATALOG_IRI],
-    [referenceCatalogPath, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [referenceCatalogPath, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-  ]);
-  assertHasCurrentWorkingFileLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    referenceCatalogPath,
-    workingLocalRelativePath,
-  );
-
-  if (versionKnopInventory) {
-    assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-      [
-        `${knopPath}/_inventory`,
-        SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-        `${knopPath}/_inventory/_history001`,
-      ],
-      [
-        `${knopPath}/_inventory/_history001`,
-        SFLO_LATEST_HISTORICAL_STATE_IRI,
-        `${knopPath}/_inventory/_history001/_s0001`,
-      ],
-    ]);
-    assertHasLiteralFacts(quads, meshBase, errorMessage, [
-      [
-        `${knopPath}/_inventory/_history001`,
-        SFLO_NEXT_STATE_ORDINAL_IRI,
-        "2",
-        XSD_NON_NEGATIVE_INTEGER_IRI,
-      ],
-    ]);
-  } else {
-    assertHasNamedNodeFacts(quads, meshBase, errorMessage, [
-      [knopPath, SFLO_HAS_RESOURCE_PAGE_IRI, `${knopPath}/index.html`],
-      [
-        `${knopPath}/_inventory`,
-        SFLO_HAS_RESOURCE_PAGE_IRI,
-        `${knopPath}/_inventory/index.html`,
-      ],
-    ]);
-  }
-
-  if (
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      referenceCatalogPath,
-      SFLO_HAS_ARTIFACT_HISTORY_IRI,
-      `${referenceCatalogPath}/_history001`,
-    )
-  ) {
-    throw new WeaveInputError(
-      `ReferenceCatalog already has explicit history for ${designatorPath}.`,
-    );
-  }
-  if (
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      referenceCatalogPath,
-      SFLO_HAS_RESOURCE_PAGE_IRI,
-      `${referenceCatalogPath}/index.html`,
-    )
-  ) {
-    throw new WeaveInputError(
-      `ReferenceCatalog already has a ResourcePage for ${designatorPath}.`,
-    );
-  }
-}
-
-function assertCurrentKnopInventoryShapeForSecondPayloadWeave(
-  meshBase: string,
-  currentKnopInventoryTurtle: string,
-  designatorPath: string,
-  payloadArtifact: PayloadWorkingArtifact,
-  nextPayloadStatePath: string,
-  options?: { knopInventoryHistoryPolicy?: SupportArtifactHistoryPolicy },
-): void {
-  const knopPath = toKnopPath(designatorPath);
-  const payloadHistoryPath = requirePayloadHistoryPath(
-    designatorPath,
-    payloadArtifact,
-  );
-  const currentPayloadStatePath = requirePayloadCurrentStatePath(
-    designatorPath,
-    payloadArtifact,
-    payloadHistoryPath,
-  );
-  const errorMessage =
-    `The current local weave slice only supports the settled second payload weave shape for ${designatorPath}.`;
-  const quads = parseWeaveShapeQuads(
-    meshBase,
-    currentKnopInventoryTurtle,
-    errorMessage,
-  );
-  const versionKnopInventory = shouldMaterializeSupportHistory(
-    options?.knopInventoryHistoryPolicy ?? "versioned",
-  );
-
-  const expectedFacts: [string, string, string][] = [
-    [knopPath, RDF_TYPE_IRI, SFLO_KNOP_IRI],
-    [knopPath, SFLO_HAS_PAYLOAD_ARTIFACT_IRI, designatorPath],
-    [designatorPath, RDF_TYPE_IRI, SFLO_PAYLOAD_ARTIFACT_IRI],
-    [designatorPath, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [designatorPath, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-    [
-      designatorPath,
-      SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-      payloadHistoryPath,
-    ],
-    [
-      payloadHistoryPath,
-      SFLO_LATEST_HISTORICAL_STATE_IRI,
-      currentPayloadStatePath,
-    ],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_KNOP_INVENTORY_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_DIGITAL_ARTIFACT_IRI],
-    [`${knopPath}/_inventory`, RDF_TYPE_IRI, SFLO_RDF_DOCUMENT_IRI],
-  ];
-  if (versionKnopInventory) {
-    expectedFacts.push(
-      [
-        `${knopPath}/_inventory`,
-        SFLO_CURRENT_ARTIFACT_HISTORY_IRI,
-        `${knopPath}/_inventory/_history001`,
-      ],
-      [
-        `${knopPath}/_inventory/_history001`,
-        SFLO_LATEST_HISTORICAL_STATE_IRI,
-        `${knopPath}/_inventory/_history001/_s0001`,
-      ],
-    );
-  }
-  assertHasNamedNodeFacts(quads, meshBase, errorMessage, expectedFacts);
-  assertHasCurrentPayloadSourceLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    designatorPath,
-    payloadArtifact,
-  );
-  assertHasCurrentWorkingFileLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    `${knopPath}/_inventory`,
-    `${knopPath}/_inventory/inventory.ttl`,
-  );
-  assertHasLiteralFacts(quads, meshBase, errorMessage, [
-    [
-      payloadHistoryPath,
-      SFLO_NEXT_STATE_ORDINAL_IRI,
-      "2",
-      XSD_NON_NEGATIVE_INTEGER_IRI,
-    ],
-  ]);
-
-  if (
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      payloadHistoryPath,
-      SFLO_HAS_HISTORICAL_STATE_IRI,
-      nextPayloadStatePath,
-    )
-  ) {
-    throw new WeaveInputError(
-      `Payload artifact already has a second explicit historical state for ${designatorPath}.`,
-    );
-  }
-  if (
-    versionKnopInventory &&
-    hasNamedNodeFact(
-      quads,
-      meshBase,
-      `${knopPath}/_inventory/_history001`,
-      SFLO_HAS_HISTORICAL_STATE_IRI,
-      `${knopPath}/_inventory/_history001/_s0002`,
-    )
-  ) {
-    throw new WeaveInputError(
-      `KnopInventory already has a second explicit historical state for ${designatorPath}.`,
-    );
-  }
 }
 
 function extractCurrentReferenceCatalogLinks(
@@ -5757,267 +4780,6 @@ function renderGenericExtractedIdentifierPage(
 `;
 }
 
-function assertHasNamedNodeFacts(
-  quads: readonly Quad[],
-  meshBase: string,
-  errorMessage: string,
-  facts: readonly (readonly [string, string, string])[],
-): void {
-  for (const [subjectValue, predicateIri, objectValue] of facts) {
-    if (
-      !hasNamedNodeFact(
-        quads,
-        meshBase,
-        subjectValue,
-        predicateIri,
-        objectValue,
-      )
-    ) {
-      throw new WeaveInputError(errorMessage);
-    }
-  }
-}
-
-function assertHasLiteralFacts(
-  quads: readonly Quad[],
-  meshBase: string,
-  errorMessage: string,
-  facts: readonly (readonly [string, string, string, string?])[],
-): void {
-  for (const [subjectValue, predicateIri, literalValue, datatypeIri] of facts) {
-    if (
-      !hasLiteralFact(
-        quads,
-        meshBase,
-        subjectValue,
-        predicateIri,
-        literalValue,
-        datatypeIri,
-      )
-    ) {
-      throw new WeaveInputError(errorMessage);
-    }
-  }
-}
-
-function assertHasCurrentPayloadSourceLocator(
-  quads: readonly Quad[],
-  meshBase: string,
-  errorMessage: string,
-  subjectValue: string,
-  payloadArtifact: PayloadWorkingArtifact,
-): void {
-  const repositorySourceFloatingLocator =
-    payloadArtifact.repositorySourceFloatingLocator;
-
-  if (repositorySourceFloatingLocator === undefined) {
-    assertHasCurrentWorkingFileLocator(
-      quads,
-      meshBase,
-      errorMessage,
-      subjectValue,
-      payloadArtifact.workingLocalRelativePath,
-    );
-    return;
-  }
-
-  if (
-    hasSubjectPredicateFact(
-      quads,
-      meshBase,
-      subjectValue,
-      SFLO_HAS_WORKING_LOCATED_FILE_IRI,
-    ) ||
-    hasSubjectPredicateFact(
-      quads,
-      meshBase,
-      subjectValue,
-      SFLO_WORKING_FILE_PATH_IRI,
-    )
-  ) {
-    throw new WeaveInputError(errorMessage);
-  }
-
-  assertHasRepositorySourceFloatingLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    subjectValue,
-    repositorySourceFloatingLocator,
-  );
-}
-
-function assertHasCurrentSourceLocator(
-  quads: readonly Quad[],
-  meshBase: string,
-  errorMessage: string,
-  subjectValue: string,
-  sourceArtifact: Pick<
-    ReferenceTargetSourcePayloadArtifact,
-    "workingLocalRelativePath" | "repositorySourceFloatingLocator"
-  >,
-): void {
-  if (sourceArtifact.repositorySourceFloatingLocator === undefined) {
-    assertHasCurrentWorkingFileLocator(
-      quads,
-      meshBase,
-      errorMessage,
-      subjectValue,
-      sourceArtifact.workingLocalRelativePath,
-    );
-    return;
-  }
-
-  if (
-    hasSubjectPredicateFact(
-      quads,
-      meshBase,
-      subjectValue,
-      SFLO_HAS_WORKING_LOCATED_FILE_IRI,
-    ) ||
-    hasSubjectPredicateFact(
-      quads,
-      meshBase,
-      subjectValue,
-      SFLO_WORKING_FILE_PATH_IRI,
-    )
-  ) {
-    throw new WeaveInputError(errorMessage);
-  }
-
-  assertHasRepositorySourceFloatingLocator(
-    quads,
-    meshBase,
-    errorMessage,
-    subjectValue,
-    sourceArtifact.repositorySourceFloatingLocator,
-  );
-}
-
-function assertHasRepositorySourceFloatingLocator(
-  quads: readonly Quad[],
-  meshBase: string,
-  errorMessage: string,
-  subjectValue: string,
-  expectedLocator: RepositorySourceFloatingLocator,
-): void {
-  const subjectIri = toAbsoluteIri(meshBase, subjectValue);
-  const locatorKeys = new Set<string>();
-
-  for (const quad of quads) {
-    if (
-      quad.subject.termType !== "NamedNode" ||
-      quad.subject.value !== subjectIri ||
-      quad.predicate.value !==
-        SFLO_HAS_REPOSITORY_SOURCE_FLOATING_LOCATOR_IRI ||
-      (quad.object.termType !== "NamedNode" &&
-        quad.object.termType !== "BlankNode")
-    ) {
-      continue;
-    }
-    locatorKeys.add(toRdfTermKey(quad.object));
-  }
-
-  if (locatorKeys.size !== 1) {
-    throw new WeaveInputError(errorMessage);
-  }
-
-  const locatorKey = locatorKeys.values().next().value!;
-  if (
-    !hasTermKeyNamedNodeFact(
-      quads,
-      locatorKey,
-      RDF_TYPE_IRI,
-      SFLO_REPOSITORY_SOURCE_FLOATING_LOCATOR_IRI,
-    )
-  ) {
-    throw new WeaveInputError(errorMessage);
-  }
-
-  const repositoryUrls = resolveUniqueLiteralValuesForTermKey(
-    quads,
-    locatorKey,
-    SFLO_SOURCE_REPOSITORY_URL_IRI,
-    errorMessage,
-  );
-  let repositoryPaths: string[];
-  try {
-    repositoryPaths = resolveUniqueLiteralValuesForTermKey(
-      quads,
-      locatorKey,
-      SFLO_SOURCE_REPOSITORY_PATH_FROM_ROOT_IRI,
-      errorMessage,
-    ).map((value) => normalizeWorkingLocalRelativePathLiteral(value));
-  } catch (error) {
-    if (error instanceof WeaveInputError) {
-      throw error;
-    }
-    throw new WeaveInputError(errorMessage);
-  }
-
-  if (
-    repositoryUrls.length !== 1 ||
-    repositoryUrls[0] !== expectedLocator.repositoryUrl ||
-    repositoryPaths.length !== 1 ||
-    repositoryPaths[0] !== expectedLocator.repositoryPathFromRoot
-  ) {
-    throw new WeaveInputError(errorMessage);
-  }
-}
-
-function assertHasCurrentWorkingFileLocator(
-  quads: readonly Quad[],
-  meshBase: string,
-  errorMessage: string,
-  subjectValue: string,
-  workingLocalRelativePath: string,
-): void {
-  const subjectIri = toAbsoluteIri(meshBase, subjectValue);
-  const values = new Set<string>();
-
-  for (const quad of quads) {
-    if (
-      quad.subject.termType !== "NamedNode" ||
-      quad.subject.value !== subjectIri
-    ) {
-      continue;
-    }
-
-    if (
-      quad.predicate.value === SFLO_HAS_WORKING_LOCATED_FILE_IRI &&
-      quad.object.termType === "NamedNode"
-    ) {
-      try {
-        values.add(
-          toMeshRelativePath(
-            meshBase,
-            quad.object.value,
-            `working file locator for ${subjectValue}`,
-          ),
-        );
-      } catch {
-        throw new WeaveInputError(errorMessage);
-      }
-      continue;
-    }
-
-    if (
-      quad.predicate.value === SFLO_WORKING_FILE_PATH_IRI &&
-      quad.object.termType === "Literal"
-    ) {
-      try {
-        values.add(normalizeWorkingLocalRelativePathLiteral(quad.object.value));
-      } catch {
-        throw new WeaveInputError(errorMessage);
-      }
-    }
-  }
-
-  if (values.size !== 1 || !values.has(workingLocalRelativePath)) {
-    throw new WeaveInputError(errorMessage);
-  }
-}
-
 function countArtifactHistoryPaths(
   meshBase: string,
   currentKnopInventoryTurtle: string,
@@ -6548,46 +5310,6 @@ function renderKnopBlockWithCarriedSupportFacts(
     workingInventoryLine,
     `${carriedLines.join("\n")}\n${workingInventoryLine}`,
   );
-}
-
-function normalizeWorkingLocalRelativePathLiteral(value: string): string {
-  const trimmed = value.trim();
-
-  if (
-    trimmed.length === 0 ||
-    trimmed.startsWith("/") ||
-    trimmed.endsWith("/") ||
-    /^[A-Za-z]:/.test(trimmed)
-  ) {
-    throw new Error("invalid workingLocalRelativePath");
-  }
-  if (
-    trimmed.includes("\\") ||
-    trimmed.includes("?") ||
-    trimmed.includes("#") ||
-    /\s/.test(trimmed)
-  ) {
-    throw new Error("invalid workingLocalRelativePath");
-  }
-
-  const normalized = pathPosix.normalize(trimmed);
-  if (normalized === "." || normalized === "..") {
-    throw new Error("invalid workingLocalRelativePath");
-  }
-
-  const segments = normalized.split("/");
-  if (segments.some((segment) => segment.length === 0)) {
-    throw new Error("invalid workingLocalRelativePath");
-  }
-
-  return normalized;
-}
-
-function usesMeshLocalWorkingLocatedFile(
-  workingLocalRelativePath: string,
-): boolean {
-  return !normalizeWorkingLocalRelativePathLiteral(workingLocalRelativePath)
-    .startsWith("../");
 }
 
 function renderCurrentWorkingFileLocator(
